Add smart time formatted text display
Also added an extra help page section warning about the possibility of
integer overflow. I could spend some time trying to patch all loose
ends, but at the end of the day it doesn't matter if setting the timer
to an obscene amount of time is going to flip, because no normal person
would have any use of setting it to a value so high anyways.