Integrate header file into source and delete it
The header file was pretty much a joke, it didn't provide any advantages
that a header normally does, simply due to the linear way in which the
code is executed in the source file.

Maybe in an alternate universe I could've written the brbtimer.c in such
a way that would allow for nice abstraction and a header file, but that
was the opposite of my goal from the very start of writing this program,
because I wanted to force myself to keep it short and not be tempted
with unnecessary enhancements. So it was only a reasonable decision to
drop the header file altogether.